requirements.txt
streamlit
pandas
numpy
faiss-cpu
torch
python-dotenv
nltk
transformers
sentence-transformers
langchain
langchain-google-genai
google-generativeai
typing-extensions
huggingface-hub
fitz
pymupdf